relocate ClubSelector
